% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem-df-build.R
\name{bdv_correct_names}
\alias{bdv_correct_names}
\title{Correct the column naming}
\usage{
bdv_correct_names(names)
}
\description{
There is an inconsistency in the column naming. Normally the structure is
almost always "<variable name>_osp_codelow" (where the latter two parts
might be missing). However, there are variables where there is a "_leg" or
"_ex" at the end. It is not always clear where the corresponding variables
come from - some of them are exactly the same with or without suffix. We will
deal with this in another function.
}
