% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem-df-build.R
\name{bdv_correct_variables}
\alias{bdv_correct_variables}
\alias{bdv_identifiers}
\alias{bdv_interval}
\alias{bdv_interval_filter}
\alias{bdv_type_C}
\title{Correct the variables}
\usage{
bdv_correct_variables(vdem_raw)

bdv_identifiers(vdem_raw)

bdv_interval(vdem_raw)

bdv_interval_filter(fxInfo_scale)

bdv_type_C(vdem_raw)
}
\description{
These functions correct the different kinds of variables.
}
\section{Functions}{
\itemize{
\item \code{bdv_identifiers}: extracts the identifier variables

\item \code{bdv_interval}: extract all interval variables.

\item \code{bdv_type_C}: Type C variables occur as "<name>",
"<name>_osp", "<name>_ord". The former two can remain as they are,
the latter is a factor and will be changed accordingly. This is very
easy by using the \code{data.frame}-method of \code{\link[haven:as_factor]{haven::as_factor()}}. The series of
dichotomous scales are regarded here as well.
}}

