% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem-mf-build.R
\name{bmv_cautionary_notes}
\alias{bmv_cautionary_notes}
\title{Include cautionary notes}
\usage{
bmv_cautionary_notes(mf_revisable)
}
\description{
The cautionary notes specify a few things that will be included as a
variable \code{fxInfo_cautionary}. The cautionary notes concering the number of
experts cannot be included here as they depend on the observation unit as
well.
}
\keyword{internal}
