% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem-mf-build.R
\name{bmv_dichotomous_series}
\alias{bmv_dichotomous_series}
\title{Create the variables from the series of dichotomous variables}
\usage{
bmv_dichotomous_series(mf_revisable)
}
\description{
These variables are given in the codebook by one variable although they
actually consist of several different variables. The new variable will have
a slightly artificial question and these variables will be the subject of
further revision. Still, these modification make the variables accessible
and do not change them so that the information from the codebook alone is
sufficient to work with them.
}
\keyword{internal}
