% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem-mf-build.R
\name{bmv_include}
\alias{bmv_include}
\title{Which names are included?}
\usage{
bmv_include(mf_revisable)
}
\description{
A few variables are not included and will thus be filtered out. This will be
marked by the keyword \code{include}. Additionally to variables that are not
present, variables from external sources will be filtered out as they have a
diverse format and I feel that it would be better to read them in from their
own source. They can be associated with the variables that are present in
this metaframe.
}
