% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem-mf-build.R
\name{bmv_spread_names}
\alias{bmv_spread_names}
\title{Spread the names}
\usage{
bmv_spread_names(mf_revisable)
}
\description{
Some names of variables actually specify several different variables, most
prominently the categorical variables. Moreover there had originally been one
typo in the name which
has been changed manually in the Codebook.
}
\keyword{internal}
