% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem-df-build.R
\name{build_df_vdem}
\alias{build_df_vdem}
\title{Build the final data frame}
\usage{
build_df_vdem(vdem_raw)
}
\description{
This function takes as input the raw data frame as read in by
\code{\link[haven:read_dta]{haven::read_dta()}}.
The building process consists of several steps that can be looked up in
detail in the sub-functions and the details to this documentation.
}
\details{
\itemize{
\item \code{\link[=bdv_correct_names]{bdv_correct_names()}} corrects inconsistencies in the column naming.
\item \code{\link[=bdv_correct_variables]{bdv_correct_variables()}} corrects the formating of the variables
where necessary.
}
}
