% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem-mf-build.R
\name{build_mf_vdem}
\alias{build_mf_vdem}
\title{Build the final metaframe}
\usage{
build_mf_vdem(mf_revisable, vdem_raw)
}
\description{
\code{mf_revisable} allows manual changes at the appropriate places in order to
facilitate quick corrections. It is then used to build the actual metaframe.
The building process consists of several steps that can be looked up in
detail in the sub-functions and the details to this documentation.
}
\details{
\itemize{
\item \code{\link[tectr:fx_read]{tectr::fx_read()}} reads in the json data files that might have been
manually changed
\item \code{\link[=bmv_spread_names]{bmv_spread_names()}} spreads the names that specify several variables
(e. g. "v2eldonate, *_osp, *_ord")
\item \code{\link[=bmv_cautionary_notes]{bmv_cautionary_notes()}} includes the cautionary notes in the codebook
\item \code{\link[=bmv_include]{bmv_include()}} marks all variables where the data release specifies that
they have not been included in version 8 as excluded (\code{include=FALSE})
\item \code{\link[=bmv_geom]{bmv_geom()}} adds the geometry arguments
}
}
