% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{df_vdem}
\alias{df_vdem}
\title{V-Dem database}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26537 rows and 2689 columns.}
\source{
\url{v-dem.net}
}
\usage{
df_vdem
}
\description{
A dataset containing the indicators of the \href{v-dem.net}{V-Dem dataset}.
}
\keyword{datasets}
