% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxGeom-OspCI.R
\docType{class}
\name{fxGeomOspCI-class}
\alias{fxGeomOspCI-class}
\alias{.fxGeomOspCI}
\title{fxGeom_class "OspCI"}
\description{
This class is intended for the osp variables. While not all of these may have
confidence intervals, there is no difference between a variable of class
"ContinuousCI" without the associated variables and "Continuous". In the name
of sparsity, we therefore only define one class.
}
