% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{vdem}
\alias{vdem}
\title{V-Dem database with metaframe}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26537 rows and 2689 columns.}
\usage{
vdem
}
\description{
This dataset combines the V-Dem database, the geocoded dataset and the
metainformation in the metaframe.
}
\keyword{datasets}
