% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{vdem_geocode}
\alias{vdem_geocode}
\title{Add a spatial column with the country polygons}
\usage{
vdem_geocode(data)
}
\arguments{
\item{data}{the data frame}
}
\description{
This functions transforms the vdem dataframe or a part of it into a
\code{\link[sf:st_sf]{sf::st_sf()}}(simple feature). It needs the columns "country_name" and
"year".
}
