% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{vdem_spatial}
\alias{vdem_spatial}
\title{Geocoded V-Dem countries}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 234 rows and 5 columns.}
\usage{
vdem_spatial
}
\description{
The coordinate polygons of all countries in the V-Dem database since 1946.
}
\keyword{datasets}
